[frontend-5] fixed errors after refactoring
