+++ frontend/src/redux/companies/companiesActions.js
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import EMPLOYERS_HH_API_URL from 'utils/constant.js';
+import {EMPLOYERS_HH_API_URL} from 'utils/constants';
 
 export const FETCH_COMPANY = 'FETCH_COMPANY';
 export const CHOOSE_COMPANY = 'CHOOSE_COMPANY';
@@ -20,7 +20,7 @@ export const chooseCompanyAction = (companyId) => {
 };
 
 export function fetchCompany(companyName) {
-    const url = EMPLOYERS_HH_API_URL + `text=${encodeURIComponent(companyName)}`;
+    const url = EMPLOYERS_HH_API_URL + `?text=${encodeURIComponent(companyName)}`;
     return (dispatch) => {
         axios.get(url)
             .then((res) => {
