+++ frontend/src/redux/competitors/competitorsActions.js
@@ -1,10 +1,11 @@
 import axios from 'axios';
-import {EMPLOYERS_HH_API_URL, BASE_URL} from 'utils/constant.js';
+import {EMPLOYERS_HH_API_URL, BASE_URL} from 'utils/constants';
 
 export const FETCH_COMPETITORS = 'FETCH_COMPETITORS';
 export const ADD_COMPETITOR = 'ADD_COMPETITOR';
 export const DELETE_COMPETITOR = 'DELETE_COMPETITOR';
 export const RESET_COMPETITORS = 'RESET_COMPETITORS';
+export const CHOOSE_COMPETITOR = 'CHOOSE_COMPETITOR';
 
 const LOGO_SIZE = 90;
 
@@ -82,7 +83,7 @@ export function addCompetitor(id, companyId) {
         "areaId": "113"
     })
     return (dispatch) => {
-        axios.get(URL_API + id).then( (el) => {
+        axios.get(EMPLOYERS_HH_API_URL + '/' + id).then( (el) => {
             let logo = el.data.logo_urls ? el.data.logo_urls[LOGO_SIZE] : null;
             dispatch(addCompetitorAction({id: el.data.id, name: el.data.name, logo: logo}));
         }
