+++ frontend/src/utils/constants.js
@@ -1,3 +1,3 @@
-export const AREAS_HH_API_URL = `https://api.hh.ru/areas`;
-export const EMPLOYERS_HH_API_URL = `https://api.hh.ru/employers`;
-export const BASE_URL = `/employer/`;
+export const AREAS_HH_API_URL = 'https://api.hh.ru/areas';
+export const EMPLOYERS_HH_API_URL = 'https://api.hh.ru/employers';
+export const BASE_URL = '/employer/';
